requirements.txt
torch>=2.0.0
torchvision>=0.15.0
nvflare-nightly @ git+https://github.com/chesterxgchen/NVFlare.git@collab_api